+++ lib/src/utils.dart
@@ -4,6 +4,7 @@ export 'utils/query_scheduler.dart';
 export 'utils/animated_switcher_utils.dart';
 export 'utils/separated_list.dart';
 export 'utils/concurrent/mutex.dart';
+export 'utils/range_joiner.dart';
 export 'utils/range_splitter.dart';
 export 'utils/transformers/debounce_stream_transformer.dart';
 export 'utils/transformers/throttle_stream_transformer.dart';